Added casts

+++ src/Compression/fuzzers/gcd_decompress_fuzzer.cpp
@@ -3,6 +3,7 @@
 
 #include <Compression/ICompressionCodec.h>
 #include <IO/BufferWithOwnMemory.h>
+#include "base/types.h"
 
 namespace DB
 {
@@ -34,7 +35,7 @@ try
     DB::Memory<> memory;
     memory.resize(output_buffer_size + codec->getAdditionalSizeAtTheEndOfBuffer());
 
-    codec->doDecompressData(reinterpret_cast<const char *>(data), size, memory.data(), output_buffer_size);
+    codec->doDecompressData(reinterpret_cast<const char *>(data), static_cast<UInt32>(size), memory.data(), static_cast<UInt32>(output_buffer_size));
 
     return 0;
 }
